+++ CHANGELOG.md
@@ -2,6 +2,7 @@
     - Added Delta TF-IDF Transformer
     - Added document frequency smoothing to TF-IDF Transformer
     - Added Unform neural net weight initializer
+    - Fixed missing probabilities in Classification Tree
 
 - 0.0.7-alpha
     - Added Support Vector Machine classifier and regressor